extract OneIdentityIdpSelectionFailureContent component
